fixed yes and no buttons
